Add BlockTable styles to support different content types
These styles specifically pair with the typed table block template in WCP PR 594. Some of them could potentially be applied as inline Tailwind classes, but I've kept them all here for easier reviewing, at least for now.

It may be better to create a mixin for some of the styles we want to get from BlockText, rather than extending it.